+++ NEWS.md
@@ -1,8 +1,10 @@
-## 7.2.0 Unreleased
+## 7.3.0 Unreleased
 
-* Adds `status` property to requests (CIRC-53)
-* Provides circulation interface 2.8 (CIRC-53)
+* Adds `systemReturnDate` property to `loans` (CIRC-81)
+* Adds `status` property to `requests` (CIRC-53)
+* Provides circulation interface 2.9 (CIRC-53, CIRC-81)
 * Requires request-storage interface 1.3 (CIRC-53)
+* Requires loan-storage interface 3.5 (CIRC-81)
 
 ## 7.1.1 2018-02-12